# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MedCard API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Visual-memory color assigned to this node
 */
export type FlowNodeTone = typeof FlowNodeTone[keyof typeof FlowNodeTone];


export const FlowNodeTone = {
  ink: 'ink',
  blue: 'blue',
  green: 'green',
  pink: 'pink',
  violet: 'violet',
  amber: 'amber',
} as const;

export interface FlowNode {
  /** Unique identifier for this node within the tree */
  id: string;
  /** The text content of this node */
  label: string;
  /**
     * Optional secondary detail shown below the label
     * @nullable
     */
  sublabel?: string | null;
  /** Child nodes branching from this node (shown side by side) */
  children?: FlowNode[];
  /** Immutable source block represented by this node */
  sourceBlockId?: string;
  /** Visual-memory color assigned to this node */
  tone?: FlowNodeTone;
}

export interface SourceBlock {
  id: string;
  /** Verbatim text copied from the user's input */
  text: string;
}

export interface SectionTrees {
  high_yield: FlowNode[];
  risk_factors: FlowNode[];
  associations: FlowNode[];
  diagnosis: FlowNode[];
  treatment: FlowNode[];
  complications: FlowNode[];
}

export type CardImageSection = typeof CardImageSection[keyof typeof CardImageSection];


export const CardImageSection = {
  main: 'main',
  high_yield: 'high_yield',
  risk_factors: 'risk_factors',
  associations: 'associations',
  diagnosis: 'diagnosis',
  treatment: 'treatment',
  complications: 'complications',
} as const;

export interface CardImage {
  id: string;
  name: string;
  /** Browser-readable image data URL */
  dataUrl: string;
  caption?: string;
  section: CardImageSection;
}

export interface SidebarSections {
  /** High-yield notes and key facts */
  high_yield: string[];
  /** Risk factors and associations */
  risk_factors: string[];
  /** Diagnostic criteria and workup findings */
  diagnosis: string[];
  /** Treatment and management */
  treatment: string[];
  /** Complications and prognosis */
  complications: string[];
}

export interface Card {
  id: number;
  /** The topic name for this card */
  topic: string;
  /** Root nodes of the pathophysiology branching tree */
  flow: FlowNode[];
  sidebar: SidebarSections;
  sectionTrees: SectionTrees;
  sourceBlocks: SourceBlock[];
  images: CardImage[];
  tags: string[];
  /** Original raw input text */
  rawText: string;
  createdAt: string;
  updatedAt: string;
}

export interface GeneratedCard {
  flow: FlowNode[];
  sidebar: SidebarSections;
  sectionTrees: SectionTrees;
  sourceBlocks: SourceBlock[];
}

export interface CardInput {
  /** @minLength 1 */
  topic: string;
  flow: FlowNode[];
  sidebar: SidebarSections;
  sectionTrees: SectionTrees;
  sourceBlocks: SourceBlock[];
  images: CardImage[];
  rawText: string;
  tags: string[];
}

export interface CardUpdate {
  /** @minLength 1 */
  topic?: string;
  flow?: FlowNode[];
  sidebar?: SidebarSections;
  sectionTrees?: SectionTrees;
  sourceBlocks?: SourceBlock[];
  images?: CardImage[];
  tags?: string[];
}

export interface GenerateCardInput {
  /**
     * Raw medical information to restructure
     * @minLength 1
     */
  rawText: string;
  /**
     * Optional topic hint
     * @nullable
     */
  topic?: string | null;
}

export type CardStatsTopTagsItem = {
  tag: string;
  count: number;
};

export interface CardStats {
  totalCards: number;
  totalTags: number;
  /** Cards created in the last 7 days */
  recentCount: number;
  topTags?: CardStatsTopTagsItem[];
}

export type ListCardsParams = {
/**
 * Search by topic or content
 */
search?: string;
/**
 * Filter by tag
 */
tag?: string;
};