# lib/api-zod/src/generated/types/card.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MedCard API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CardImage } from './cardImage';
import type { FlowNode } from './flowNode';
import type { SectionTrees } from './sectionTrees';
import type { SidebarSections } from './sidebarSections';
import type { SourceBlock } from './sourceBlock';

export interface Card {
  id: number;
  /** The topic name for this card */
  topic: string;
  /** Root nodes of the pathophysiology branching tree */
  flow: FlowNode[];
  sidebar: SidebarSections;
  sectionTrees: SectionTrees;
  sourceBlocks: SourceBlock[];
  images: CardImage[];
  tags: string[];
  /** Original raw input text */
  rawText: string;
  createdAt: Date;
  updatedAt: Date;
}