# lib/api-zod/src/generated/types/cardInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MedCard API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CardImage } from './cardImage';
import type { FlowNode } from './flowNode';
import type { SectionTrees } from './sectionTrees';
import type { SidebarSections } from './sidebarSections';
import type { SourceBlock } from './sourceBlock';

export interface CardInput {
  /** @minLength 1 */
  topic: string;
  flow: FlowNode[];
  sidebar: SidebarSections;
  sectionTrees: SectionTrees;
  sourceBlocks: SourceBlock[];
  images: CardImage[];
  rawText: string;
  tags: string[];
}