# lib/api-zod/src/generated/types/flowNode.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MedCard API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FlowNodeTone } from './flowNodeTone';

export interface FlowNode {
  /** Unique identifier for this node within the tree */
  id: string;
  /** The text content of this node */
  label: string;
  /**
     * Optional secondary detail shown below the label
     * @nullable
     */
  sublabel?: string | null;
  /** Child nodes branching from this node (shown side by side) */
  children?: FlowNode[];
  /** Immutable source block represented by this node */
  sourceBlockId?: string;
  /** Visual-memory color assigned to this node */
  tone?: FlowNodeTone;
}