# lib/api-zod/src/generated/types/sectionTrees.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MedCard API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FlowNode } from './flowNode';

export interface SectionTrees {
  high_yield: FlowNode[];
  risk_factors: FlowNode[];
  associations: FlowNode[];
  diagnosis: FlowNode[];
  treatment: FlowNode[];
  complications: FlowNode[];
}